Move `OutboundStopSimPayload` to `stateful::message`
