+++ packages/engine/lib/stateful/src/message/mod.rs
@@ -1,7 +1,6 @@
-mod payload;
+pub mod payload;
 mod schema;
 
-pub use self::{
-    payload::GenericPayload,
-    schema::{MessageSchema, MESSAGE_ARROW_FIELDS, MESSAGE_BATCH_SCHEMA, MESSAGE_COLUMN_NAME},
+pub use self::schema::{
+    MessageSchema, MESSAGE_ARROW_FIELDS, MESSAGE_BATCH_SCHEMA, MESSAGE_COLUMN_NAME,
 };
